feat(#302): Refactor term handling by adding separate classes for to_float and to_time
