+++ test/factbase/terms/test_to_float.rb
@@ -5,19 +5,15 @@
 
 require_relative '../../test__helper'
 require_relative '../../../lib/factbase/term'
+require_relative '../../../lib/factbase/terms/to_float'
 
-# Math test.
+# Test for 'to_float' term.
 # Author:: Yegor Bugayenko (yegor256@gmail.com)
 # Copyright:: Copyright (c) 2024-2025 Yegor Bugayenko
 # License:: MIT
-class TestCasting < Factbase::Test
+class TestToFloat < Factbase::Test
   def test_to_float
-    t = Factbase::Term.new(:to_float, [[3.14, 'hello']])
+    t = Factbase::ToFloat.new([[3.14, 'hello']])
     assert_equal('Float', t.evaluate(fact, [], Factbase.new).class.to_s)
-  end
-
-  def test_to_time
-    t = Factbase::Term.new(:to_time, [%w[2023-01-01 hello]])
-    assert_equal('Time', t.evaluate(fact, [], Factbase.new).class.to_s)
   end
 end
